começa a separar componentes da interface
